agents/media_performance_agent/requirements.txt
# Core A2A and Agent SDK packages
a2a-sdk==0.2.9
toolbox-core
google_adk==1.4.2
google-cloud-aiplatform[agent_engines,adk]==1.97.0
mcp==1.8.0

# Web framework and server
fastapi==0.115.13
starlette==0.46.2
uvicorn==0.34.3
sse-starlette==2.3.6

# HTTP clients and async utilities
httpx==0.28.1
httpx-sse==0.4.1
asyncclick==8.1.8.0
click==8.2.1
requests==2.32.4

# Data validation and configuration
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
python-dotenv==1.1.1

# Web scraping and parsing
beautifulsoup4==4.12.3
html2text==2024.2.26

# Database drivers
pg8000==1.31.2

# Utilities
deprecated